+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [0.1.2] - 2019-01-15
 ### Added
 - Send and then close alert to non-Safari browsers to try to avoid focus problems
 
@@ -19,4 +19,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Initial Code
 
+[0.1.2]: https://github.com/squarebracket/karma-selenium-grid-launcher/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/squarebracket/karma-selenium-grid-launcher/compare/0.1.0...0.1.1